app/requirements.txt
grpcio==1.75.1
grpcio-tools==1.75.1
protobuf==6.33.0
typing_extensions==4.15.0
